Refactor recent calls subcomponents
